- All Voxel formats now with y as up axis - Meshers can now work without the need of materials and textures - Rewrite of the meshbuilder

+++ lib/src/Formats/Implementations/GoxelFormat.hpp
@@ -67,8 +67,8 @@ namespace VoxelOptimizer
 
             struct Block
             {
-                Math::Vec3f Pos;
-                size_t Index;
+                Math::Vec3i Pos;
+                int Index;
             };
 
             struct Layer
